Restored test for `EnumeratesValues::getArrayableItems`
