+++ tests/Support/Common.php
@@ -2,9 +2,12 @@
 
 namespace Illuminate\Tests\Support;
 
+use ArrayIterator;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Contracts\Support\Jsonable;
+use IteratorAggregate;
 use JsonSerializable;
+use Traversable;
 
 class TestArrayableObject implements Arrayable
 {
@@ -27,5 +30,33 @@ class TestJsonSerializeObject implements JsonSerializable
     public function jsonSerialize(): array
     {
         return ['foo' => 'bar'];
+    }
+}
+
+class TestJsonSerializeWithScalarValueObject implements JsonSerializable
+{
+    public function jsonSerialize(): string
+    {
+        return 'foo';
+    }
+}
+
+class TestTraversableAndJsonSerializableObject implements IteratorAggregate, JsonSerializable
+{
+    public $items;
+
+    public function __construct($items)
+    {
+        $this->items = $items;
+    }
+
+    public function getIterator(): Traversable
+    {
+        return new ArrayIterator($this->items);
+    }
+
+    public function jsonSerialize(): array
+    {
+        return json_decode(json_encode($this->items), true);
     }
 }
