+++ tests/Support/SupportCollectionTest.php
@@ -19,6 +19,7 @@ use JsonSerializable;
 use Mockery as m;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
+use ReflectionClass;
 use stdClass;
 use Symfony\Component\VarDumper\VarDumper;
 use UnexpectedValueException;
@@ -592,6 +593,46 @@ class SupportCollectionTest extends TestCase
         })->values();
 
         $this->assertSame([3, 3, 4, 4], $data->all());
+    }
+
+    /**
+     * @dataProvider collectionClassProvider
+     */
+    public function testGetArrayableItems($collection)
+    {
+        $data = new $collection;
+
+        $class = new ReflectionClass($collection);
+        $method = $class->getMethod('getArrayableItems');
+
+        $items = new TestArrayableObject;
+        $array = $method->invokeArgs($data, [$items]);
+        $this->assertSame(['foo' => 'bar'], $array);
+
+        $items = new TestJsonableObject;
+        $array = $method->invokeArgs($data, [$items]);
+        $this->assertSame(['foo' => 'bar'], $array);
+
+        $items = new TestJsonSerializeObject;
+        $array = $method->invokeArgs($data, [$items]);
+        $this->assertSame(['foo' => 'bar'], $array);
+
+        $items = new TestJsonSerializeWithScalarValueObject;
+        $array = $method->invokeArgs($data, [$items]);
+        $this->assertSame(['foo'], $array);
+
+        $subject = [new stdClass, new stdClass];
+        $items = new TestTraversableAndJsonSerializableObject($subject);
+        $array = $method->invokeArgs($data, [$items]);
+        $this->assertSame($subject, $array);
+
+        $items = new $collection(['foo' => 'bar']);
+        $array = $method->invokeArgs($data, [$items]);
+        $this->assertSame(['foo' => 'bar'], $array);
+
+        $items = ['foo' => 'bar'];
+        $array = $method->invokeArgs($data, [$items]);
+        $this->assertSame(['foo' => 'bar'], $array);
     }
 
     /**
